Notify new frames via callbacks
Currently, a frame is available to the consumer as soon as it is pushed
by the producer (which can detect if the previous frame is skipped).

Notify the new frames (and frame skipped) via callbacks instead.

This pave the way to add (optional) buffering, which will introduce a
delay between the time when the frame is produced and the time it is
available to be consumed.